refactor: pair ~56 ABI-identical signatures with the binary (close coverage holes)
Wave 2: these functions were byte-identical on the ARM32 cross-build but
their MANGLED names differed (const&/qualifier, or int<->long width), so
asm-verify -- which pairs on exact mangled name -- never diffed their
bodies. They were silent coverage holes, not "verified MATCH". Aligning
the port signatures to the binary makes asm-verify pair + diff them, so
future body drift surfaces (the class of gap this session repeatedly
found hiding real bugs).

All mechanical, no behavior change (ARM32 passes const& and by-value
aggregates the same; long==int==4B on both ARM32 and Windows x64):
- const& -> by-value where the binary passes by value (Colour, Vec3,
  Delegate0, ...).
- const method-qualifier added/removed to match the binary.
- int/unsigned int -> long/unsigned long width where the binary uses long.
- A handful of const& -> non-const& (Math::ClosestPointOnLine/LineIntersect,
  FancyBakedStringBuild) with local-copy/const_cast call-site fixups.
- Game::Init/MortarGame::Init -> (int, const char**) (coupled override).

Skipped (deferred): 6 SmartPtr/AsciiString by-value cases (host copy
cost), the 7 ES2 Draw-refactor + 5 templated/std findings (intentional /
noise), and the 2 D0/D2 virtual-destructor-variant mismatches
(Game::~Game, NetworkManager::~NetworkManager -- vtable-layout question,
need re-analyst). Host + full-game build green.

+++ src/engine/input/Touch.cpp
@@ -114,7 +114,7 @@ void Touch::Update(float dt) {
 // Binary @ 0x00195690 -- __UpdateInternal.
 // Push TEvnt to ring; on overflow: Update(0.0f) then retry.
 // isActive: true=press OR move, false=release (matches binary param 'b').
-void Touch::__UpdateInternal(uint32_t extId, bool isActive, float x, float y, float t) {
+void Touch::__UpdateInternal(unsigned long extId, bool isActive, float x, float y, float t) {
     int next = eventBuffer.m_eventTail + 1;
     if (next - eventBuffer.m_eventHead >= MAX_EVENTS) {
         Update(0.0f);
@@ -136,7 +136,7 @@ void Touch::__UpdateInternal(uint32_t extId, bool isActive, float x, float y, fl
 // Apply event to states2: scan for matching extId, then claim free slot if new press.
 // isActive: true=press OR move, false=release.
 // Binary @ 0x00195314 -- free slot predicate is extId==0, NOT phase>=1.
-void Touch::___UpdateInternal(uint32_t extId, bool isActive, float x, float y) {
+void Touch::___UpdateInternal(unsigned long extId, bool isActive, float x, float y) {
 #ifdef FN_DEBUG_TOUCH
     LOG_DEBUG("TOUCH", "drain extId=%u active=%d x=%g y=%g\n",
              extId, (int)isActive, x, y);
@@ -182,7 +182,7 @@ void Touch::___UpdateInternal(uint32_t extId, bool isActive, float x, float y) {
 
 // Binary @ 0x00195424 -- FindTouch(uint touchId).
 // Linear scan states1; return slot index or -1.
-int Touch::FindTouch(uint32_t touchId) {
+int Touch::FindTouch(unsigned long touchId) {
     for (int i = 0; i < MAX_SLOTS; i++) {
         if (states1[i].touchId == touchId) return i;
     }
@@ -209,7 +209,7 @@ uint32_t Touch::GetMostRecentTouch() {
 // ASM-spec v1.6.1 Touch::GetTouchPos @0x002429d4: (uint, float&, float&).
 // Writes currX/Y of matching slot. Returns 1 if active (phase < 1), 0 if not.
 // Binary leaves *x/*y UNTOUCHED on miss -- do not zero them.
-int Touch::GetTouchPos(uint32_t touchId, float& x, float& y) {
+int Touch::GetTouchPos(unsigned long touchId, float& x, float& y) {
     int slot = FindTouch(touchId);
     if (slot < 0) return 0;
     x = states1[slot].currX;
@@ -219,7 +219,7 @@ int Touch::GetTouchPos(uint32_t touchId, float& x, float& y) {
 
 // ASM-spec v1.6.1 Touch::GetTouchDelta @0x00242a20: (uint, float&, float&).
 // Writes currX-prevX/dy if phase >= 0, else 0.0f. Returns 1 if active.
-int Touch::GetTouchDelta(uint32_t touchId, float& dx, float& dy) {
+int Touch::GetTouchDelta(unsigned long touchId, float& dx, float& dy) {
     int slot = FindTouch(touchId);
     if (slot < 0 || states1[slot].phase >= 1) { dx = 0.0f; dy = 0.0f; return 0; }
     if (states1[slot].phase < 0) {
